+++ src/sxpdb.cpp
@@ -115,8 +115,8 @@ SEXP add_val_origin_(SEXP sxpdb, SEXP val,
     }
   }
   catch(std::exception& e) {
-    Rf_error("Error adding value from package %s, function %s and argument %s, with call id %lu, into the database: %s\n",
-             package_name, function_name, argument_name, call_id, e.what());
+    Rf_error("Error adding value from package %s, function %s and argument %s, with call id %llu, into the database: %s\n",
+             package_name, function_name, argument_name, (unsigned long long) call_id, e.what());
   }
   
   if(!db->is_quiet()) {
